requirements.txt
chromadb==0.5.0
datasets==2.19.1
gpt4all==2.6.0
joblib==1.4.2
langchain==0.2.6
langchain_chroma==0.1.2
langchain_community==0.2.6
langchain_core==0.2.11
langchain_text_splitters==0.2.2
numpy==2.0.0
pandas==2.2.2
peft==0.11.1
PyYAML==6.0.1
torch==2.2.2
tqdm==4.66.2
transformers==4.41.1
trl==0.8.6
